+++ Firestore/core/src/remote/datastore.cc
@@ -16,6 +16,7 @@
 
 #include "Firestore/core/src/remote/datastore.h"
 
+#include <unordered_map>
 #include <unordered_set>
 #include <utility>
 
@@ -281,7 +282,9 @@ void Datastore::RunAggregateQueryWithCredentials(
     const core::Query& query,
     const std::vector<AggregateField>& aggregates,
     api::AggregateQueryCallback&& callback) {
-  absl::flat_hash_map<std::string, std::string> aliasMap;
+  // TODO(b/443765747) Revert back to absl::flat_hash_map after the absl version
+  // is upgraded to later than 20250127.0
+  std::unordered_map<std::string, std::string> aliasMap;
   grpc::ByteBuffer message =
       MakeByteBuffer(datastore_serializer_.EncodeAggregateQueryRequest(
           query, aggregates, aliasMap));