+++ FirebaseAI/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+- [feature] Added support for the URL context tool, which allows the model to access content
+  from provided public web URLs to inform and enhance its responses. (#15221)
+- [changed] Using Firebase AI Logic with the Gemini Developer API is now Generally Available (GA).
+- [changed] Using Firebase AI Logic with the Imagen generation APIs is now Generally Available (GA).
+
 # 12.3.0
 - [feature] Added support for the Code Execution tool, which enables the model
   to generate and run code to perform complex tasks like solving mathematical